Refactor imports and update project structure
Refactor import paths across several files to improve project organization. Moved types from `src/types/contractValues.ts` to `src/contracts-api/contractValues.ts` to better align with their usage context. This change streamlines module imports and clarifies file responsibilities.
